Moved the SourceModelParser
